Add jaeger configuration
