+++ multichain-api-gateway/config/base.env
@@ -3,3 +3,6 @@
 #MULTICHAIN_API_GATEWAY__BLOCKSCOUT__CONCURRENT_REQUESTS=10
 #MULTICHAIN_API_GATEWAY__BLOCKSCOUT__REQUEST_TIMEOUT=60
 #MULTICHAIN_API_GATEWAY__BLOCKSCOUT__INSTANCES="eth/mainnet;xdai/mainnet;xdai/testnet;eth/testnet"
+
+#MULTICHAIN_API_GATEWAY__JAEGER__ENABLED=false
+#MULTICHAIN_API_GATEWAY__JAEGER__AGENT_ENDPOINT=localhost:6831
