+++ multichain-api-gateway/config/base.toml
@@ -4,11 +4,13 @@ addr = "0.0.0.0:8044"
 [blockscout]
 base_url = "https://blockscout.com"
 concurrent_requests = 10
-
 instances = [
     "eth/mainnet",
     "xdai/testnet",
     "xdai/mainnet",
 ]
-
 request_timeout = 60
+
+[jaeger]
+enabled = false
+agent_endpoint = "localhost:6831"
